adding functionality to upload images to products, and create pdf of order invoices and stream it to user.

+++ src/controllers/admin.js
@@ -1,6 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
-const mongoose = require('mongoose');
+const { deleteFile } = require('../utils/file');
 
 module.exports.getAddProduct = (req, res, _next) => {
   res.render('admin/edit-product', {
@@ -14,20 +14,37 @@ module.exports.getAddProduct = (req, res, _next) => {
 };
 
 module.exports.postAddProduct = (req, res, next) => {
-  const { title, imageUrl, description, price } = req.body;
+  const { title, description, price } = req.body;
+  const userId = req.user._id;
+  const image = req.file;
   const errors = validationResult(req);
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      product: { title, description, price },
+      hasError: true,
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: [],
+    });
+  }
+
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
       editing: false,
-      product: { title, imageUrl, description, price },
+      product: { title, description, price },
       hasError: true,
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
-  const userId = req.user._id;
+
+  const imageUrl = image.path;
+
   const product = new Product({
     title,
     price,
@@ -87,28 +104,35 @@ module.exports.getEditProduct = (req, res, next) => {
 };
 
 module.exports.postEditProduct = (req, res, next) => {
-  const { title, imageUrl, description, price, productId } = req.body;
+  const { title, description, price, productId } = req.body;
+  const image = req.file;
+
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Edit Product',
+      path: '/admin/add-product',
+      editing: true,
+      product: { _id: productId, title, price, description },
+      hasError: true,
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
+    });
+  }
+
   Product.findById(productId)
     .then((product) => {
       if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
       }
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(422).render('admin/edit-product', {
-          pageTitle: 'Edit Product',
-          path: '/admin/add-product',
-          editing: true,
-          product: { _id: productId, title, price, description, imageUrl },
-          hasError: true,
-          errorMessage: errors.array()[0].msg,
-          validationErrors: errors.array(),
-        });
-      }
+
       product.title = title;
       product.price = price;
       product.description = description;
-      product.imageUrl = imageUrl;
+      if (image) {
+        deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then((result) => {
         console.log('Product Updated!');
         res.redirect('/admin/products');
@@ -123,7 +147,14 @@ module.exports.postEditProduct = (req, res, next) => {
 
 module.exports.postDeleteProduct = (req, res, next) => {
   const id = req.body.productId;
-  Product.deleteOne({ _id: id, userId: req.user._id })
+  Product.findById(id)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found!'));
+      }
+      deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: id, userId: req.user._id });
+    })
     .then((result) => {
       res.redirect('/admin/products');
     })
